+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## v1.2.1 - 2024-30-04
+* Make description field in APIs and Products optional.
+* Fix logic error in product policy resource introduced in v1.2.0.
+* Using `try()` instead of conditional `can()` in arguments.
+
 ## v1.2.0 - 2024-04-26
 * Added option to fallback to default policy filename (`policy.xml`), if custom policy filename doesn't exist. This can be useful if custom policy filename is used to differentiate between runtime environments, but in some APIs or products the policy should be the same in all runtime environments. In this case, you then only need to configure one `policy.xml`.